refactor(ReservationConfirmActivity): 할인 적용 금액을 알기 위한 유즈케이스를 활용해서 도메인과 의존성을 더 감소시킴
