+++ app/src/main/java/woowacourse/movie/ui/confirm/ReservationConfirmActivity.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.ui.confirm
 
 import android.os.Bundle
 import android.widget.TextView
-import com.example.domain.discountPolicy.DiscountPolicy
+import com.example.domain.usecase.DiscountApplyUseCase
 import woowacourse.movie.R
 import woowacourse.movie.model.ReservationState
 import woowacourse.movie.model.mapper.asDomain
@@ -14,17 +14,17 @@ import java.text.DecimalFormat
 import java.time.format.DateTimeFormatter
 
 class ReservationConfirmActivity : BackKeyActionBarActivity() {
+    private val discountApplyUseCase = DiscountApplyUseCase()
+
     private val titleTextView: TextView by lazy { findViewById(R.id.reservation_title) }
     private val dateTextView: TextView by lazy { findViewById(R.id.reservation_date) }
     private val moneyTextView: TextView by lazy { findViewById(R.id.reservation_money) }
     private val reservationCountTextView: TextView by lazy { findViewById(R.id.reservation_count) }
 
-    private val discountCalculator: DiscountPolicy by lazy { DiscountPolicy() }
     override fun onCreateView(savedInstanceState: Bundle?) {
         setContentView(R.layout.activity_reservation_confirm)
-        val reservationRes =
-            intent.getParcelableExtraCompat<ReservationState>(KEY_RESERVATION)
-                ?: return keyError(KEY_RESERVATION)
+        val reservationRes = intent.getParcelableExtraCompat<ReservationState>(KEY_RESERVATION)
+            ?: return keyError(KEY_RESERVATION)
         setInitReservationData(reservationRes)
     }
 
@@ -33,15 +33,15 @@ class ReservationConfirmActivity : BackKeyActionBarActivity() {
     ) {
         titleTextView.text = reservationState.movieState.title
         dateTextView.text = reservationState.dateTime.format(DATE_TIME_FORMATTER)
-        moneyTextView.text = formattingMoney(reservationState)
         reservationCountTextView.text =
             getString(R.string.person_count_text, reservationState.countState.value)
+        setDiscountApplyMoney(reservationState)
     }
 
-    private fun formattingMoney(reservationState: ReservationState): String {
-        val money = discountCalculator.discount(reservationState.asDomain()).value
-        return DECIMAL_FORMATTER.format(money)
-    }
+    private fun setDiscountApplyMoney(reservationState: ReservationState) =
+        discountApplyUseCase(reservationState.asDomain()) {
+            moneyTextView.text = DECIMAL_FORMATTER.format(it.value)
+        }
 
     companion object {
         private val DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern("yyyy.M.d HH:mm")
